requirements.txt
matplotlib
nibabel
numpy
pillow
scipy
torch
torchvision
torchinfo
tqdm